requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
mediapipe
scikit-learn
joblib
opencv-python-headless
